[Client] Improve keyboard movement
The movement of a character was not triggered in case the key to move
was just pressed a short time. This issue was caused by the delay timer
that is in place to allow proper movement with two keys. This issue is
now fixed. In case the delay timer did not expire, the movement is
triggered once the key is released.