[Client] Catching problems at calculating the line of sight
The bresenham implementation caused a error to be thrown in case the
line got too long. While this indicates a error it shouldn't send the
client to hell. The most likely reason is that the line is calculated
for two character that are too far apart from each other. This error is
now caught and handled properly.