[Build] Fixing versions of resources
The resources are now treated like external dependencies with fixed
version numbers that need to be updated by hand. Appears to be
currently the best way to push the updates in the proper way.

With the dynamic dependency on the the GIT version the main problem is
that with each release of the applications a new version of the
resources needs to be released as well to fit the version number. No
matter if the resources were changed or not.