[Client] Prevent double-clicks from leaking through GUI windows
For double-clicks the first click is consumed twice. Once as a single click and
once as the first click of the double-click. This fix makes sure, that the second
click is not passed on to the listener because of that.

Probably this is not the best fix possible, but anything else would require a major
rewrite of how input is handled.

This fixes mantis issue #9553.