[Client] refs #11407 Disable launch aggressive
The launch aggressive option forces java to apply aggressive and
potentially unsafe optimizations during the runtime. Tests showed that
the there was no performance improvement due to this change. How ever a
user reported that he tracked by freezes of the client to this option.

For that reason this option will be disabled.